requirements.txt
openai
langchain
unstructured
streamlit
python-dotenv
PyPDF2
faiss-cpu

-e .